add route search with opts

+++ models/Movie.go
@@ -1,13 +1,13 @@
 package models
 
 type Movie struct {
-	Id int
-	Title string
+	Id          int64
+	Title       string
 	Description string
-	Duration string
-	Artists string
-	Genre string
-	WatchURL string
-	Vote int64
-	ViewCount int64
+	Duration    string
+	Artists     string
+	Genre       string
+	WatchURL    string
+	Vote        int64
+	ViewCount   int64
 }
